ImageSEOStream/pyproject.toml
[project]
name = "imageseo-pro"
version = "2.0.0"
description = "Advanced AI Image Analyzer & SEO Optimizer with GPT-4o Vision"
authors = [
    {name = "ImageSEO Pro Team", email = "support@imageseo.pro"}
]
readme = "README.md"
requires-python = ">=3.11"
keywords = ["ai", "image-analysis", "seo", "openai", "gpt-4o", "image-optimization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "numpy>=2.2.6",
    "openai>=1.99.9",
    "opencv-contrib-python>=4.12.0.88",
    "opencv-python>=4.12.0.88",
    "pathlib>=1.0.1",
    "pillow>=11.3.0",
    "streamlit>=1.48.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/imageseo-pro/imageseo-pro"
Documentation = "https://docs.imageseo.pro"
Repository = "https://github.com/imageseo-pro/imageseo-pro"
Issues = "https://github.com/imageseo-pro/imageseo-pro/issues"

[project.scripts]
imageseo-pro = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
